Introduce _convert_to_iterable_if_not_none
In a lot of places we get either one value or list-like object
with values inside. For simplifying the code if we have one value
we make a tuple with one element out of it and then iterate over it
as usual. This helper function makes a lot of similar code go away